Refactor : 장소 업로드 부분 이미지 캐러셀 부분 Refactor
이미지 캐러셀 부분 MVVM 모델 적용
