+++ client/project-there/src/components/placeUpload/CustomCarousel.jsx
@@ -1,41 +1,15 @@
-import React, { useState, useEffect, useContext } from 'react';
+import React, { useState, useEffect } from 'react';
 import 'react-responsive-carousel/lib/styles/carousel.min.css';
 import { Carousel } from 'react-responsive-carousel';
-import Resizer from 'react-image-file-resizer';
-import exifr from 'exifr';
 import styled from 'styled-components';
 import CustomDropZone from './CustomDropZone';
 import { usePlaceInfoActions, usePlaceInfoValue } from '../../contexts/PlaceInfoContext';
 import CarouselItem from './CarouselItem';
-
-const resizeFileToBase64 = (file) => new Promise((resolve) => {
-  Resizer.imageFileResizer(
-    file,
-    600,
-    600,
-    'JPEG',
-    100,
-    0,
-    (uri) => {
-      resolve(uri);
-    },
-    'base64',
-  );
-});
-const resizeFileToFile = (file) => new Promise((resolve) => {
-  Resizer.imageFileResizer(
-    file,
-    600,
-    600,
-    'JPEG',
-    100,
-    0,
-    (uri) => {
-      resolve(uri);
-    },
-    'file',
-  );
-});
+import {
+  transformImagesToBase64,
+  resizeImagesUpdateImageData,
+  extractImageInfoAndUpdateData,
+} from '../../lib/hooks/usePlaceUpload';
 
 const StyledCarousel = styled(Carousel)`
   width: 100%;
@@ -55,10 +29,11 @@ const StyledCarousel = styled(Carousel)`
 export default function CustomCarousel({ containerRef }) {
   const [acceptedImages, setAcceptedImages] = useState([]);
   const [, setRejectedImages] = useState([]);
-  const [resizedImages, setResizedImages] = useState([]);
+  const [transformedImages, setTransformedImages] = useState([]);
   const { updateData } = usePlaceInfoActions();
   const [containerHeight, setContainerHeight] = useState(300);
   const placeInfo = usePlaceInfoValue();
+  const [isUpdate, setIsUpdate] = useState(true);
 
   const addAcceptedImages = (preImages, curImages) => {
     setAcceptedImages([...preImages, ...curImages]);
@@ -68,46 +43,17 @@ export default function CustomCarousel({ containerRef }) {
   };
 
   useEffect(() => {
-    if (placeInfo?.images?.length !== 0) {
+    if (placeInfo?.images?.length !== 0 && isUpdate) {
       setAcceptedImages(placeInfo?.images);
     }
-  }, [placeInfo?.images?.length]);
+  }, [placeInfo?.images?.length, isUpdate]);
 
   useEffect(() => {
-    Promise.all(acceptedImages?.map((image) => resizeFileToBase64(image))).then((result) => {
-      setResizedImages(result);
-    });
-    Promise.all(acceptedImages?.map((image) => resizeFileToFile(image))).then((result) => {
-      updateData('images', result);
-    });
-    Promise.all(acceptedImages?.map((image) => exifr.parse(image)))
-      .then((result) => {
-        let placeVisitedTime = new Date();
-        const presentTime = placeVisitedTime;
-        const imageTakenLocations = [];
-        const findDuplicate = [];
-        result.forEach((info) => {
-          if (info) {
-            const { CreateDate, latitude, longitude } = info;
-            if (CreateDate) {
-              placeVisitedTime = CreateDate < placeVisitedTime ? CreateDate : placeVisitedTime;
-            }
-            if (latitude && longitude) {
-              if (!findDuplicate.includes(`${latitude}${longitude}`)) {
-                findDuplicate.push(`${latitude}${longitude}`);
-                imageTakenLocations.push({ lat: latitude, lng: longitude });
-              }
-            }
-          }
-        });
-        if (imageTakenLocations.length !== 0) {
-          updateData('imageTakenLocations', imageTakenLocations);
-        }
-        if (presentTime !== placeVisitedTime) {
-          updateData('placeTime', placeVisitedTime);
-        }
-      });
-  }, [acceptedImages]);
+    setIsUpdate(false);
+    transformImagesToBase64(acceptedImages, setTransformedImages);
+    resizeImagesUpdateImageData(acceptedImages, updateData);
+    extractImageInfoAndUpdateData(acceptedImages, updateData);
+  }, [acceptedImages?.length]);
 
   useEffect(() => {
     if (containerRef) {
@@ -121,7 +67,7 @@ export default function CustomCarousel({ containerRef }) {
       infiniteLoop
       showThumbs={false}
     >
-      {resizedImages?.map((imageUrl, index) => (
+      {transformedImages?.map((imageUrl, index) => (
         <CarouselItem
           key={imageUrl}
           src={imageUrl}
